Find the Longest Substring Without Repeating Characters.cpp
You are given a string S, and your task is to find the length of the longest substring that contains no repeating characters. A substring is a contiguous block of characters in the string.
In this problem, you need to find the length of the longest substring where all the characters are unique. The substring can be formed by starting at any position in the string, but it must not contain any repeated characters.

Input :
* A string S, where 1≤∣S∣≤105(length of string).
* Input: S = "abcabcbb"

Output :
An integer representing the length of the longest substring without repeating characters.
Output: 3