The Coin Change Problem.cpp
You are given an integer array coins[] of size n, where each element represents the denomination of a coin. You are also given an integer amount, representing the total amount of money. The task is to find the minimum number of coins required to make up the given amount.
If the amount cannot be formed by any combination of the coins, return -1.
You can assume that you have an infinite supply of each denomination.

Input :
* An integer array coins[] where each element represents the value of a coin.
* An integer amount representing the total amount of money.
* coins = [1, 2, 5], amount = 11

Output :
* Return the minimum number of coins needed to make up the amount.
* If the amount cannot be formed by any combination of coins, return -1.
* Output - 3